requirements.txt
alembic==1.13.1
auth==0.5.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.0
ecdsa==0.18.0
eventlet==0.35.1
falcon==3.1.3
flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
flask-sqlalchemy==3.1.1
future==0.18.3
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jose==1.0.0
Mako==1.3.2
MarkupSafe==2.1.5
mongoengine==0.27.0
packaging==23.2
psycopg2==2.9.9
pyasn1==0.5.1
pycryptodome==3.3.1
pymongo==4.6.1
python-dotenv==1.0.1
python-jose==3.3.0
python-jose-cryptodome==1.3.2
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.27
typing-extensions==4.9.0
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.17.0
